feat: 단위 테스트 적용을 위한 예외처리 코드 추가

+++ yaxim/src/main/java/com/yaxim/user/entity/Users.java
@@ -27,8 +27,8 @@ public class Users {
     @Setter
     private boolean active;
 
-    public Users(int id, String name, String email) {
-        this.id = (long) id;
+    public Users(Long id, String name, String email) {
+        this.id = id;
         this.name = name;
         this.email = email;
     }
